Remove need for unused TokenFactoryQuery, use shared stake hooks
